实现问题回复

+++ src/main/java/life/wmm/commulity/community/controller/PublishController.java
@@ -36,7 +36,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model
     ) {
@@ -61,7 +61,7 @@ public class PublishController {
         question.setTitle(title);
         question.setDescription(description);
         question.setTag(tag);
-        question.setCreator(Integer.valueOf(user.getId()));
+        question.setCreator(user.getId());
 
         question.setId(id);
         questionService.createOrUpdate(question);
@@ -70,7 +70,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,Model model){
+    public String edit(@PathVariable(name = "id") Long id,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
         model.addAttribute("description", question.getDescription());
